+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Changelog
 * 1.0.0-next (@release_date@)
 
  * added `BindingState::DUPLICATE` to output of `BindingState::all()`
+ * renamed `ManagerEvents` to `PuliEvents` and moved to 
+   `Puli\RepositoryManager\Api\Event\` namespace
 
 * 1.0.0-beta2 (2015-01-27)